+++ geocode_failures_output.txt
@@ -1,0 +1,116 @@
+Geocode failure count: 113
+Geocode failed addresses:
+ Cherry just east of 23
+19tg and Sansom 5min
+Test street
+Market just west of 18th
+Market just west of 20th
+Walnut just east of 17th
+22 just north of market street
+Yes
+19 street between market and chestnut
+Xy
+
+15 th street just south of chancellor
+Market between 21 and 20
+23rd just north of walnut street
+23rd street just north of walnut
+Market just east of 20th
+Outside f 3609 market
+Market street a little west of 19th
+15th close to chancellor
+Yes
+38 at Samson
+17th street right before walnut.
+15th a little north of walnut.
+15 th a little south of walnut
+15th just south of walnut street
+15th just north of walnut.
+15 & Samson
+15th just north of walnut
+4600 block of Hazel ace
+Samson at spruce
+Spruce and Samson
+
+18th just north of market
+15 and market (15 st station)
+18th just north of market
+22nd and Market southwest corner septa station.
+38 between spruce and chestnut
+19 between Sansom and walnut
+19 between Sansom and walnut
+Alley between 1845 walnut and 135 s 19th off walnut
+Spruce between 13 and juniper
+1033 Germantown (piazza)
+19th just past market
+3 between race and market
+Spruce between broad and 13
+14 th and Bach street
+14 th street a little past locust
+16 between spruce and locust
+Market between 22 and 21
+15 and walnut on 15 th
+Walnut street between 17 and 18
+Lia Toyota dealership
+12 th between market and chestnut
+19 th and Samson
+Delivery truck idling behind Haymarket on Crackerbarrel Alley
+19th between Samson and walnut
+20th to Samson
+22nd just north of market
+22nd between chestnut and market
+22 nd between chestnut and market
+22nd between chestnut and market
+Samson between 16 and 15th
+22nd and market trolley stop SW corner.
+Samson between 16 and 17
+Swanson street between 16 and 17.
+Market close to 19th
+Sansom between 6&7
+Market between 22nd and 21st.
+5 th street past market
+19 between walnut and Sansom
+Sansom between 6&7
+Front and Lombard St. Parking lot on east side if Front
+5 past market on east side
+5th street past market on east side
+Chestnut between 7&8
+Sansom between 18&19
+Yes
+Yes
+
+Yes
+Walnut between broad and 13th
+Yes
+Walnut between 18&19
+Arch between 10&11
+Arch between 10&11
+Arch between 10&11
+Arch between 10&11
+Arch between 10&11
+Arch between 9&10
+Arch between 9&10
+6th between arch & race
+22nd and market trolley stop NW corner
+Locust between 20 & 21
+22nd and Market southwest corner trolley stop.
+Sansom between 6 & 7
+13th and movarian
+Market between 7 & 8
+Market between 7 & 8
+Walnut between 16 & 17
+Broad between locust and spruce
+2nd and Naudain, then 100 block of Front St.
+Market between 5 & 4
+Market between 5 & 4
+Market between 5 & 4
+Market between 5 & 4
+Market between 5 & 4
+12 between chestnut and market
+market street between 22 and 21
+5 just north of market
+6 between market and chestnut
+Sansom between 6&7
+Walnut btwn 18&19
+Spruce between 17&18
+9 but idling when I approached
