+++ manage.py
@@ -4,6 +4,7 @@ from app import create_app, db
 from app.models import User, Role, Agency, Permission, IncidentReport
 from flask.ext.script import Manager, Shell
 from flask.ext.migrate import Migrate, MigrateCommand
+from app.utils import parse_to_db
 
 # Import settings from .env file. Must define FLASK_CONFIG
 if os.path.exists('.env'):
@@ -108,6 +109,17 @@ def setup_dev():
     db.session.add(general)
 
     db.session.commit()
+
+
+@manager.option('-f',
+                '--filename',
+                default='poll244.csv',
+                type=str,
+                help='Filename of csv to parse',
+                dest='filename')
+def parse_csv(filename):
+    """Parses the given csv file into the database."""
+    parse_to_db(db, filename, True)
 
 
 @manager.command
